chore: hoist `app` into root dir & intro `ThemeProvider` to web app
