+++ apps/web/src/providers/index.ts
@@ -1,0 +1,1 @@
+export * from './ThemeProvider';
